+++ src/extension.ts
@@ -205,6 +205,15 @@ export module BackgroundPhiColors
     const indentModeObject = Object.freeze({ "none": null, "light": null, "smart": null, "full": null, });
     const tokenModeObject = Object.freeze ({ "none": null, "light": null, "smart": null, "full": null, });
     const activeScopeObject = Object.freeze ({ "editor": null, "document": null, "window": null, });
+    const laneObject = Object.freeze
+    (
+        {
+            "left": vscode.OverviewRulerLane.Left,
+            "center": vscode.OverviewRulerLane.Center,
+            "right": vscode.OverviewRulerLane.Right,
+            "full": vscode.OverviewRulerLane.Full,
+        }
+    );
 
     const enabled = new Config("enabled", true);
     const enabledPanels = new Config("enabledPanels", false);
@@ -228,8 +237,11 @@ export module BackgroundPhiColors
     const trailingSpacesEnabled = new Config("trailingSpacesEnabled", true);
     const symbolEnabled = new Config("symbolEnabled", false);
     const showIndentErrorInOverviewRuler = new Config("showIndentErrorInOverviewRuler", true);
+    const indentErrorInOverviewRulerLane = new Config<keyof typeof laneObject>("indentErrorInOverviewRulerLane", "left", makeEnumValidator(Object.keys(laneObject)));
     const showActiveTokenInOverviewRuler = new Config("showActiveTokenInOverviewRuler", true);
+    const activeTokenInOverviewRulerLane = new Config<keyof typeof laneObject>("activeTokenInOverviewRulerLane", "center", makeEnumValidator(Object.keys(laneObject)));
     const showTrailingSpacesErrorInOverviewRuler = new Config("showTrailingSpacesErrorInOverviewRuler", true);
+    const trailingSpacesErrorInOverviewRulerLane = new Config<keyof typeof laneObject>("trailingSpacesErrorInOverviewRulerLane", "right", makeEnumValidator(Object.keys(laneObject)));
     const spacesAlpha =new Config("spacesAlpha", 0x11, undefined, 0x00, 0xFF);
     const spacesActiveAlpha =new Config("spacesActiveAlpha", 0x33, undefined, 0x00, 0xFF);
     const spacesErrorAlpha =new Config("spacesErrorAlpha", 0x88, undefined, 0x00, 0xFF);
@@ -297,7 +309,7 @@ export module BackgroundPhiColors
             base: hslaCache.get(spaceErrorColor.get(lang)),
             hue: 0,
             alpha: spacesErrorAlpha.get(lang),
-            overviewRulerLane: showIndentErrorInOverviewRuler.get(lang) ? vscode.OverviewRulerLane.Left: undefined,
+            overviewRulerLane: showIndentErrorInOverviewRuler.get(lang) ? laneObject[indentErrorInOverviewRulerLane.get(lang)]: undefined,
         }
     );
     const makeTrailingSpacesErrorDecorationParam = (lang: string): DecorationParam =>
@@ -307,7 +319,7 @@ export module BackgroundPhiColors
             base: hslaCache.get(spaceErrorColor.get(lang)),
             hue: 0,
             alpha: spacesErrorAlpha.get(lang),
-            overviewRulerLane: showTrailingSpacesErrorInOverviewRuler.get(lang) ? vscode.OverviewRulerLane.Right: undefined,
+            overviewRulerLane: showTrailingSpacesErrorInOverviewRuler.get(lang) ? laneObject[trailingSpacesErrorInOverviewRulerLane.get(lang)]: undefined,
         }
     );
     let decorations: { [decorationParamJson: string]: { decorator: vscode.TextEditorDecorationType, rangesOrOptions: vscode.Range[] } } = { };
@@ -595,8 +607,11 @@ export module BackgroundPhiColors
             trailingSpacesEnabled,
             symbolEnabled,
             showIndentErrorInOverviewRuler,
+            indentErrorInOverviewRulerLane,
             showActiveTokenInOverviewRuler,
+            activeTokenInOverviewRulerLane,
             showTrailingSpacesErrorInOverviewRuler,
+            trailingSpacesErrorInOverviewRulerLane,
             spacesAlpha,
             spacesActiveAlpha,
             spacesErrorAlpha,
@@ -912,7 +927,7 @@ export module BackgroundPhiColors
                                                         tokenBaseColor,
                                                         hash(i),
                                                         tokenActiveAlpha,
-                                                        showActiveTokenInOverviewRuler.get(lang) ? vscode.OverviewRulerLane.Center: undefined,
+                                                        showActiveTokenInOverviewRuler.get(lang) ? laneObject[activeTokenInOverviewRulerLane.get(lang)]: undefined,
                                                     )
                                                 }
                                             )
@@ -1526,7 +1541,7 @@ export module BackgroundPhiColors
                         tokenBaseColor,
                         i.specificColor || hash(i.token),
                         i.isActive ? tokenActiveAlpha: tokenAlpha,
-                        i.isActive && showActiveTokenInOverviewRuler.get(lang) ? vscode.OverviewRulerLane.Center: undefined,
+                        i.isActive && showActiveTokenInOverviewRuler.get(lang) ? laneObject[activeTokenInOverviewRulerLane.get(lang)]: undefined,
                     )
                 }
             )